requirements.txt
Django==5.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
drf-yasg==1.21.7
Pillow>=10.4.0
